Remove Patient, Practitioner, and PractitionerRole
Remove profiles for Patient, Practitioner, and PractitionerRole. We should rather refer to the Finnish FHIR Base Profiles, and not repeat the profiles here.

These profiles impose many constraints on cardinalities. Especially setting the maximum cardinality of elements to 0. We should avoid that in implementation guides with many potential use cases.