Relax maximum cardinalities
In most cases, it is not appropriate to limit the upper limit of cardinalities. It enhances interoperability if systems providing information are able to provide several codings of the element. Also, in some use cases you may need more instances of an element.

The receiving system is free to ignore the elements it does not support (as long as they are not modifier elements, of course).

We may well want to create profiles where we say that an element with a certain coding needs to be present. But we should not mandate that no other elements or codings are allowed to be present as well. This should be done by slicing.