Remove reference to codeserver
The most recent ig publisher expects this to be a proper FHIR value set, not just a link to a HTML page.

We could adopt the value set and introduce it in FHIR format in this IG. Or we can wait for HL7 Finland to get a proper FHIR terminology server set up. Clarifying requirements for the latter are in the work plan for this year, and at least both THL and Duodecim have expressed interest in working on this.