Make publish more robust
This takes more time to build, but is more robust.
We want to empty the temp folder in order to not make the resulting zip archive too big. Everything in temp seems to be copied there.
Also, make sure to run _genonce at least one, so that the version is updated.